+++ superfields/src/main/java/org/vaadin/miki/superfields/dates/SuperDatePicker.java
@@ -1,19 +1,30 @@
 package org.vaadin.miki.superfields.dates;
 
 import com.vaadin.flow.component.AbstractField;
+import com.vaadin.flow.component.AttachEvent;
+import com.vaadin.flow.component.ClientCallable;
+import com.vaadin.flow.component.DetachEvent;
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.datepicker.DatePicker;
 import com.vaadin.flow.component.dependency.JsModule;
-import org.vaadin.miki.markers.HasLabel;
-import org.vaadin.miki.markers.HasLocale;
-import org.vaadin.miki.markers.HasPlaceholder;
+import com.vaadin.flow.shared.Registration;
+import org.vaadin.miki.events.text.TextSelectionListener;
+import org.vaadin.miki.events.text.TextSelectionNotifier;
+import org.vaadin.miki.markers.CanReceiveSelectionEventsFromClient;
+import org.vaadin.miki.markers.CanSelectText;
+import org.vaadin.miki.markers.WithDatePatternMixin;
 import org.vaadin.miki.markers.WithIdMixin;
 import org.vaadin.miki.markers.WithLabelMixin;
 import org.vaadin.miki.markers.WithLocaleMixin;
 import org.vaadin.miki.markers.WithPlaceholderMixin;
+import org.vaadin.miki.markers.WithReceivingSelectionEventsFromClientMixin;
 import org.vaadin.miki.markers.WithValueMixin;
+import org.vaadin.miki.shared.dates.DatePattern;
+import org.vaadin.miki.shared.text.TextSelectionDelegate;
 
 import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
 import java.util.Locale;
 
 /**
@@ -23,14 +34,18 @@ import java.util.Locale;
  */
 @JsModule("./super-date-picker.js")
 @Tag("super-date-picker")
+@SuppressWarnings("squid:S110") // there is no way to reduce the number of parent classes
 public class SuperDatePicker extends DatePicker
-        implements HasLocale, HasLabel, HasPlaceholder, HasDatePattern,
+        implements CanSelectText, CanReceiveSelectionEventsFromClient, WithReceivingSelectionEventsFromClientMixin<SuperDatePicker>,
+                   TextSelectionNotifier<SuperDatePicker>,
                    WithLocaleMixin<SuperDatePicker>, WithLabelMixin<SuperDatePicker>,
                    WithPlaceholderMixin<SuperDatePicker>, WithDatePatternMixin<SuperDatePicker>,
                    WithValueMixin<AbstractField.ComponentValueChangeEvent<DatePicker, LocalDate>, LocalDate, SuperDatePicker>,
                    WithIdMixin<SuperDatePicker> {
 
-    private final DatePatternHelper<SuperDatePicker> delegate = new DatePatternHelper<>(this);
+    private final DatePatternDelegate<SuperDatePicker> datePatternDelegate = new DatePatternDelegate<>(this);
+
+    private final TextSelectionDelegate<SuperDatePicker> textSelectionDelegate = new TextSelectionDelegate<>(this, this.getEventBus(), this::getFormattedValue);
 
     private DatePattern datePattern;
 
@@ -87,8 +102,8 @@ public class SuperDatePicker extends DatePicker
     public final void setLocale(Locale locale) {
         // there is a call for setting locale from the superclass' constructor
         // and when that happens, the field is not yet initialised
-        if(this.delegate != null) {
-            this.delegate.initPatternSetting();
+        if(this.datePatternDelegate != null) {
+            this.datePatternDelegate.initPatternSetting();
             SuperDatePickerI18nHelper.updateI18N(locale, this::getI18n, this::setI18n);
         }
         super.setLocale(locale);
@@ -97,11 +112,70 @@ public class SuperDatePicker extends DatePicker
     @Override
     public void setDatePattern(DatePattern datePattern) {
         this.datePattern = datePattern;
-        this.delegate.updateClientSidePattern();
+        this.datePatternDelegate.updateClientSidePattern();
     }
 
     @Override
     public DatePattern getDatePattern() {
         return this.datePattern;
+    }
+
+    @Override
+    protected void onAttach(AttachEvent attachEvent) {
+        this.textSelectionDelegate.onAttach(attachEvent, super::onAttach);
+    }
+
+    @Override
+    protected void onDetach(DetachEvent detachEvent) {
+        this.textSelectionDelegate.onDetach(detachEvent, super::onDetach);
+    }
+
+    @ClientCallable
+    private void selectionChanged(int start, int end, String selection) {
+        this.textSelectionDelegate.fireTextSelectionEvent(true, start, end, selection);
+    }
+
+    @Override
+    public boolean isReceivingSelectionEventsFromClient() {
+        return this.textSelectionDelegate.isReceivingSelectionEventsFromClient();
+    }
+
+    @Override
+    public void setReceivingSelectionEventsFromClient(boolean receivingSelectionEventsFromClient) {
+        this.textSelectionDelegate.setReceivingSelectionEventsFromClient(receivingSelectionEventsFromClient);
+    }
+
+    /**
+     * Returns the current value formatted with current locale or pattern.
+     * @return Current date, formatted. Will return {@code null} if the current date is {@code null}.
+     */
+    public String getFormattedValue() {
+        final LocalDate value = this.getValue();
+        if(value == null)
+            return null;
+        else if(this.getDatePattern() == null)
+            return value.format(DateTimeFormatter.ofLocalizedDate(FormatStyle.SHORT).withLocale(this.getLocale()));
+        else
+            return this.datePatternDelegate.formatDate(value);
+    }
+
+    @Override
+    public void selectAll() {
+        this.textSelectionDelegate.selectAll();
+    }
+
+    @Override
+    public void selectNone() {
+        this.textSelectionDelegate.selectNone();
+    }
+
+    @Override
+    public void select(int from, int to) {
+        this.textSelectionDelegate.select(from, to);
+    }
+
+    @Override
+    public Registration addTextSelectionListener(TextSelectionListener<SuperDatePicker> listener) {
+        return this.textSelectionDelegate.addTextSelectionListener(listener);
     }
 }