+++ superfields/src/main/java/org/vaadin/miki/superfields/dates/SuperDateTimePicker.java
@@ -6,12 +6,12 @@ import com.vaadin.flow.component.datepicker.DatePicker;
 import com.vaadin.flow.component.datetimepicker.DateTimePicker;
 import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.timepicker.TimePicker;
-import org.vaadin.miki.markers.HasLabel;
-import org.vaadin.miki.markers.HasLocale;
+import org.vaadin.miki.markers.WithDatePatternMixin;
 import org.vaadin.miki.markers.WithIdMixin;
 import org.vaadin.miki.markers.WithLabelMixin;
 import org.vaadin.miki.markers.WithLocaleMixin;
 import org.vaadin.miki.markers.WithValueMixin;
+import org.vaadin.miki.shared.dates.DatePattern;
 import org.vaadin.miki.util.ReflectTools;
 
 import java.time.LocalDateTime;
@@ -25,9 +25,9 @@ import java.util.Optional;
  */
 @JsModule("./super-date-time-picker.js")
 @Tag("super-date-time-picker")
+@SuppressWarnings("squid:S110") // there is no way to reduce the number of parent classes
 public class SuperDateTimePicker extends DateTimePicker
-        implements HasLocale, HasLabel, HasDatePattern,
-                   WithLocaleMixin<SuperDateTimePicker>, WithLabelMixin<SuperDateTimePicker>,
+        implements WithLocaleMixin<SuperDateTimePicker>, WithLabelMixin<SuperDateTimePicker>,
                    WithDatePatternMixin<SuperDateTimePicker>, WithIdMixin<SuperDateTimePicker>,
                    WithValueMixin<AbstractField.ComponentValueChangeEvent<DateTimePicker, LocalDateTime>, LocalDateTime, SuperDateTimePicker> {
 
@@ -36,7 +36,7 @@ public class SuperDateTimePicker extends DateTimePicker
     private static final String INTERNAL_DATE_PICKER_FIELD_NAME = "datePicker";
     private static final String INTERNAL_TIME_PICKER_FIELD_NAME = "timePicker";
 
-    private final DatePatternHelper<SuperDateTimePicker> delegate = new DatePatternHelper<>(this);
+    private final DatePatternDelegate<SuperDateTimePicker> delegate = new DatePatternDelegate<>(this);
 
     private DatePattern datePattern;