+++ superfields/release-notes.md
@@ -1,3 +1,21 @@
+# 0.7.0 - SuperTextField
+## New features and enhancements
+* \#122 - [SuperTextField, SuperTextArea and text selection API](https://api.github.com/repos/vaadin-miki/super-fields/issues/122)
+* \#123 - [Add text selection API to existing components](https://api.github.com/repos/vaadin-miki/super-fields/issues/123)
+* \#135 - [SuperTabs should have an option to wrap tabs](https://api.github.com/repos/vaadin-miki/super-fields/issues/135)
+* \#141 - [Server-side date formatting for DatePattern](https://api.github.com/repos/vaadin-miki/super-fields/issues/141)
+* \#152 - [Allow overriding default value in number fields](https://api.github.com/repos/vaadin-miki/super-fields/issues/152)
+## Changes to API
+* \#123 - [Add text selection API to existing components](https://api.github.com/repos/vaadin-miki/super-fields/issues/123)
+* \#147 - [HasId marker interface](https://api.github.com/repos/vaadin-miki/super-fields/issues/147)
+* \#171 - [DatePattern should be moved to a shared package](https://github.com/vaadin-miki/super-fields/issues/171)
+## Bug fixes
+* \#132 - [No deployment to Heroku](https://api.github.com/repos/vaadin-miki/super-fields/issues/132)
+* \#136 - [setReadOnly has no effect on number fields](https://api.github.com/repos/vaadin-miki/super-fields/issues/136)
+* \#137 - [UnloadObserver.onDetach causes TypeError](https://api.github.com/repos/vaadin-miki/super-fields/issues/137)
+* \#146 - [SuperTabs are difficult to style](https://api.github.com/repos/vaadin-miki/super-fields/issues/146)
+* \#154 - [Number fields do not trigger focus/blur events](https://api.github.com/repos/vaadin-miki/super-fields/issues/154)
+
 # 0.6.2 - Vaadin 14.2 compatibility
 ## New features and enhancements
 * \#124 - [Update Vaadin dependencies to 14.2](https://api.github.com/repos/vaadin-miki/super-fields/issues/124)